src/requirements.txt
certifi==2019.6.16
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
pandas==0.24.0
pyparsing==2.4.2
pyserial==3.4
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
virtualenv==16.7.2
virtualenv-clone==0.5.3
